Combine jsp view controller
